backend/requirements.txt
bafser==2.2
Flask-MonitoringDashboard==4.0.4
python-dotenv
requests
gunicorn